PostInsights refactor to utils/postInsights
